add model logic for reserving tables and
this includes a `reserve_table` method and a `avaiable_table` method
`avaiable_table` is a query method that checks the database for tables based on request
capacity. Tables are produced in the seed file. It returns a distinct table and the first available. That table is no longer available after a successful reservation.
This avoids duplicates and overbooking